+++ asdf/asdf.py
@@ -419,13 +419,11 @@ class AsdfFile(versioning.VersionedMixin):
         return None
 
     @classmethod
-    def _open_impl(cls, self, fd, uri=None, mode='r',
+    def _open_asdf(cls, self, fd, uri=None, mode='r',
                    validate_checksums=False,
                    do_not_fill_defaults=False,
                    _get_yaml_content=False):
-        if not is_asdf_file(fd):
-            raise ValueError("Does not appear to be a ASDF file.")
-
+        """Attempt to populate AsdfFile data from file-like object"""
         fd = generic_io.get_file(fd, mode=mode, uri=uri)
         self._fd = fd
         header_line = fd.read_until(b'\r?\n', 2, "newline", include=True)
@@ -485,10 +483,37 @@ class AsdfFile(versioning.VersionedMixin):
         return self
 
     @classmethod
+    def _open_impl(cls, self, fd, uri=None, mode='r',
+                   validate_checksums=False,
+                   do_not_fill_defaults=False,
+                   _get_yaml_content=False,
+                  accept_asdf_in_fits=True):
+        """Attempt to open file-like object as either AsdfFile or AsdfInFits"""
+        if not is_asdf_file(fd):
+            msg = "Input object does not appear to be ASDF file"
+            if accept_asdf_in_fits:
+                try:
+                    # TODO: this feels a bit circular, try to clean up. Also
+                    # this introduces another dependency on astropy which may
+                    # not be desireable.
+                    from . import fits_embed
+                    return fits_embed.AsdfInFits.open(fd, uri=uri,
+                                validate_checksums=validate_checksums,
+                                extensions=self._extensions)
+                except ValueError:
+                    msg += " or FITS with an ASDF extension"
+            raise ValueError(msg)
+        return cls._open_asdf(self, fd, uri=uri, mode=mode,
+                validate_checksums=validate_checksums,
+                do_not_fill_defaults=do_not_fill_defaults,
+                _get_yaml_content=_get_yaml_content)
+
+    @classmethod
     def open(cls, fd, uri=None, mode='r',
              validate_checksums=False,
              extensions=None,
-             do_not_fill_defaults=False):
+             do_not_fill_defaults=False,
+             accept_asdf_in_fits=True):
         """
         Open an existing ASDF file.
 
@@ -518,6 +543,11 @@ class AsdfFile(versioning.VersionedMixin):
         do_not_fill_defaults : bool, optional
             When `True`, do not fill in missing default values.
 
+        accept_asdf_in_fits : bool, optional
+            When `True`, try to automatically process FITS files with ASDF
+            extensions. If backwards compatibility with old behavior is
+            required, set this flag to `False`.
+
         Returns
         -------
         asdffile : AsdfFile
@@ -528,7 +558,8 @@ class AsdfFile(versioning.VersionedMixin):
         return cls._open_impl(
             self, fd, uri=uri, mode=mode,
             validate_checksums=validate_checksums,
-            do_not_fill_defaults=do_not_fill_defaults)
+            do_not_fill_defaults=do_not_fill_defaults,
+            accept_asdf_in_fits=accept_asdf_in_fits)
 
     def _write_tree(self, tree, fd, pad_blocks):
         fd.write(constants.ASDF_MAGIC)
@@ -983,14 +1014,13 @@ def is_asdf_file(fd):
         return True
     elif isinstance(fd, generic_io.GenericFile):
         pass
-    elif isinstance(fd, io.IOBase):
+    else:
         try:
             fd = generic_io.get_file(fd, mode='r', uri=None)
+            if not isinstance(fd, io.IOBase):
+                to_close = True
         except ValueError:
             return False
-    else:
-        to_close = True
-        fd = generic_io.get_file(fd, mode='r', uri=None)
     asdf_magic = fd.read(5)
     if fd.seekable():
         fd.seek(0)
@@ -998,5 +1028,4 @@ def is_asdf_file(fd):
         fd.close()
     if asdf_magic == constants.ASDF_MAGIC:
         return True
-    else:
-        return False
+    return False